Fix some random Error Prone warnings.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=170241825

+++ android/guava/src/com/google/common/util/concurrent/SmoothRateLimiter.java
@@ -253,7 +253,7 @@ abstract class SmoothRateLimiter extends RateLimiter {
         permitsToTake -= permitsAboveThresholdToTake;
       }
       // measuring the integral on the left part of the function (the horizontal line)
-      micros += (stableIntervalMicros * permitsToTake);
+      micros += (long) (stableIntervalMicros * permitsToTake);
       return micros;
     }
